+++ src/FromStringOptions.ts
@@ -1,0 +1,10 @@
+
+/**
+ * Options configuring behavior of fromString
+ */
+export interface FromStringOptions {
+
+    ellipsis?: string;
+
+    componentPrefix?: string;
+}