+++ src/internal/MicrogrammarSpecParser.ts
@@ -3,10 +3,23 @@ import { Concat, toMatchingLogic } from "../matchers/Concat";
 import { Literal } from "../Primitives";
 
 import { WhiteSpaceHandler } from "../Config";
+import { FromStringOptions } from "../FromStringOptions";
 import { isPatternMatch } from "../PatternMatch";
 import { Break } from "./Break";
+import { CompleteFromStringOptions, completeWithDefaults } from "./CompleteFromStringOptions";
 import { exactMatch } from "./ExactMatch";
 import { MicrogrammarSpec, specGrammar } from "./SpecGrammar";
+
+/**
+ * Convenient function to create a microgrammar from a spec within another grammar
+ * @param spec string spec
+ * @param components
+ * @param options
+ * @returns {Concat}
+ */
+export function fromString(spec: string, components: object = {}, options: FromStringOptions = {}): Concat {
+    return new MicrogrammarSpecParser().fromString(spec, components, options);
+}
 
 /**
  * Parses microgrammars expressed as strings.
@@ -15,24 +28,45 @@ export class MicrogrammarSpecParser {
 
     private anonFieldCount = 0;
 
-    public fromString(spec: string, elements: object = {}): Concat {
-        const match = exactMatch<MicrogrammarSpec>(specGrammar, spec);
+    public fromString(spec: string, components: object, options: FromStringOptions): Concat {
+        const optionsToUse = completeWithDefaults(options);
+        spec = this.preprocess(spec, optionsToUse);
+        const match = exactMatch<MicrogrammarSpec>(specGrammar(optionsToUse), spec);
         if (!isPatternMatch(match)) {
             throw new Error(`Unable to parse microgrammar: ${spec}`);
         }
         const matcherSequence1 = this.definitionSpecsFromMicrogrammarSpec(match,
-            (elements as WhiteSpaceHandler).$consumeWhiteSpaceBetweenTokens !== false);
-        const matcherSequence2 = this.populateSpecifiedElements(elements, matcherSequence1);
+            (components as WhiteSpaceHandler).$consumeWhiteSpaceBetweenTokens !== false);
+        const matcherSequence2 = this.populateSpecifiedElements(components, matcherSequence1);
         const matcherSequence3 = this.inferUnspecifiedElements(matcherSequence2);
         const definitions = this.definitionsFromSpecs(spec, matcherSequence3);
         const concat = new Concat(definitions);
         // Copy config to Concat
-        for (const key in elements) {
+        for (const key in components) {
             if (key.charAt(0) === "$") {
-                concat[key] = elements[key];
+                concat[key] = components[key];
             }
         }
         return concat;
+    }
+
+    /**
+     * Given a spec, replace all the DiscardToken instances with a named,
+     * but unbound, matcher spec
+     * @param spec spec to preprocess before parsing
+     * @param optionsToUse options
+     * @returns {string}
+     */
+    private preprocess(spec: string, optionsToUse: CompleteFromStringOptions): string {
+        const split = spec.split(optionsToUse.ellipsis);
+        let joined = "";
+        for (let i = 0; i < split.length; i++) {
+            if (i > 0) {
+                joined += `${optionsToUse.componentPrefix}{_discard${i}}`;
+            }
+            joined += split[i];
+        }
+        return joined;
     }
 
     private definitionSpecsFromMicrogrammarSpec(match: MicrogrammarSpec, consumeWhiteSpaceBetweenTokens: boolean): DefinitionSpec[] {
@@ -89,8 +123,8 @@ export class MicrogrammarSpecParser {
     }
 
     private definitionsFromSpecs(id: string, definitionSpecs: SatisfiedDefinitionSpec[]): object {
-        const definitions = {$id: id};
-        definitionSpecs.forEach( t  => {
+        const definitions = { $id: id };
+        definitionSpecs.forEach(t => {
             if (isAnonymous(t)) {
                 this.addAnonymousToDefinitions(definitions, t.anonymous);
             }