+++ src/internal/SpecGrammar.ts
@@ -2,23 +2,29 @@ import { Concat } from "../matchers/Concat";
 import { RestOfInput } from "../matchers/skip/Skip";
 import { Rep } from "../Rep";
 import { Break } from "./Break";
+import { CompleteFromStringOptions } from "./CompleteFromStringOptions";
 
-const elementReference = new Concat({
-    $id: "component",
-    _start: "${",
-    elementName: /[a-zA-Z0-9_]+/,
-    _end: "}",
-});
+function componentReference(fso: CompleteFromStringOptions) {
+    return new Concat({
+        $id: "component",
+        _start: fso.componentPrefix + "{",
+        elementName: /[a-zA-Z0-9_]+/,
+        _end: "}",
+    });
+}
 
-export const specGrammar = new Concat({
-    $id: "spec",
-    these: new Rep({
-        $id: "literal, then component",
-        literal: new Break(elementReference),
-        element: elementReference,
-    }),
-    trailing: RestOfInput,
-});
+export function specGrammar(fso: CompleteFromStringOptions) {
+    const componentRef = componentReference(fso);
+    return new Concat({
+        $id: "spec",
+        these: new Rep({
+            $id: "literal, then component",
+            literal: new Break(componentRef),
+            element: componentRef,
+        }),
+        trailing: RestOfInput,
+    });
+}
 
 export interface MicrogrammarSpec {
     these: Array<{ literal: string, element: { elementName: string } }>;